// src/pinecone-generated-ts-fetch/admin/models/Project.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Pinecone Admin API
 * Provides an API for managing a Pinecone organization and its resources, including projects, API keys, organization users and invites, service accounts, and role bindings. 
 *
 * The version of the OpenAPI document: 2026-07
 * Contact: support@pinecone.io
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { exists, mapValues } from '../runtime';
/**
 * The details of a project.
 * @export
 * @interface Project
 */
export interface Project {
    /**
     * The unique ID of the project.
     * @type {string}
     * @memberof Project
     */
    id: string;
    /**
     * The name of the project.
     * @type {string}
     * @memberof Project
     */
    name: string;
    /**
     * The maximum number of Pods that can be created in the project.
     * @type {number}
     * @memberof Project
     */
    maxPods: number;
    /**
     * Whether to force encryption with a customer-managed encryption key (CMEK).
     * @type {boolean}
     * @memberof Project
     */
    forceEncryptionWithCmek: boolean;
    /**
     * The unique ID of the organization that the project belongs to.
     * @type {string}
     * @memberof Project
     */
    organizationId: string;
    /**
     * The date and time when the project was created.
     * @type {Date}
     * @memberof Project
     */
    createdAt?: Date;
}

/**
 * Check if a given object implements the Project interface.
 */
export function instanceOfProject(value: object): boolean {
    let isInstance = true;
    isInstance = isInstance && "id" in value;
    isInstance = isInstance && "name" in value;
    isInstance = isInstance && "maxPods" in value;
    isInstance = isInstance && "forceEncryptionWithCmek" in value;
    isInstance = isInstance && "organizationId" in value;

    return isInstance;
}

export function ProjectFromJSON(json: any): Project {
    return ProjectFromJSONTyped(json, false);
}

export function ProjectFromJSONTyped(json: any, ignoreDiscriminator: boolean): Project {
    if ((json === undefined) || (json === null)) {
        return json;
    }
    return {
        
        'id': json['id'],
        'name': json['name'],
        'maxPods': json['max_pods'],
        'forceEncryptionWithCmek': json['force_encryption_with_cmek'],
        'organizationId': json['organization_id'],
        'createdAt': !exists(json, 'created_at') ? undefined : (new Date(json['created_at'])),
    };
}

export function ProjectToJSON(value?: Project | null): any {
    if (value === undefined) {
        return undefined;
    }
    if (value === null) {
        return null;
    }
    return {
        
        'id': value.id,
        'name': value.name,
        'max_pods': value.maxPods,
        'force_encryption_with_cmek': value.forceEncryptionWithCmek,
        'organization_id': value.organizationId,
        'created_at': value.createdAt === undefined ? undefined : (value.createdAt.toISOString()),
    };
}